refactor(tooling): separate code size gate support
