+++ scripts/code-size-gate-baseline.ts
@@ -1,0 +1,31 @@
+import type { CodeSizeBaselineEntry } from './code-size-gate-types.js';
+
+export const CODE_SIZE_BASELINE: readonly CodeSizeBaselineEntry[] =
+  Object.freeze([
+    { path: 'examples/docs/stories.ts', lines: 4558 },
+    { path: 'examples/docs/app.ts', lines: 3306 },
+    { path: 'packages/bijou-tui/src/app-frame.ts', lines: 2847 },
+    { path: 'packages/bijou/src/core/components/table.ts', lines: 990 },
+    { path: 'packages/bijou-tui/src/runtime-engine.ts', lines: 975 },
+    { path: 'packages/bijou/src/core/render/differ.ts', lines: 901 },
+    { path: 'packages/bijou/src/core/ui-scene-ir.ts', lines: 889 },
+    { path: 'packages/bijou-tui/src/app-frame-overlays.ts', lines: 877 },
+    { path: 'examples/docs/i18n-debt.ts', lines: 831 },
+    { path: 'packages/bijou/src/ports/surface.ts', lines: 823 },
+    { path: 'packages/bijou-tui/src/app-frame-render.ts', lines: 763 },
+    { path: 'scripts/pr-review-status.ts', lines: 754 },
+    { path: 'examples/image-viewer/main.ts', lines: 740 },
+    { path: 'examples/notifications/main.ts', lines: 725 },
+    { path: 'examples/_shared/canonical-app.ts', lines: 724 },
+    { path: 'packages/bijou-tui-app/src/index.ts', lines: 692 },
+    { path: 'packages/bijou/src/core/components/dag-render.ts', lines: 692 },
+    { path: 'packages/bijou-tui/src/driver.ts', lines: 679 },
+    { path: 'examples/perf-gradient/main.ts', lines: 676 },
+    { path: 'packages/bijou-tui/src/runtime.ts', lines: 669 },
+    { path: 'packages/bijou-tui/src/timeline.ts', lines: 632 },
+    { path: 'examples/docs/storybook-app.ts', lines: 607 },
+    { path: 'packages/bijou-i18n/src/runtime.ts', lines: 565 },
+    { path: 'scripts/smoke-all-examples-lib.ts', lines: 558 },
+    { path: 'packages/bijou-tui/src/app-frame-actions.ts', lines: 557 },
+    { path: 'packages/bijou-tui/src/eventbus.ts', lines: 534 },
+  ]);
